sndh: Fix uint32_t to snprintf
lib/psgplay/sndh.c: In function ‘sndh_frames_subtag’:
lib/psgplay/sndh.c:288:60: warning: format ‘%u’ expects argument of type ‘unsigned int’, but argument 4 has type ‘uint32_t’ {aka ‘long unsigned int’} [-Wformat=]
  288 |         snprintf(cursor->buffer, sizeof(cursor->buffer), "%u", t);
      |                                                           ~^   ~
      |                                                            |   |
      |                                                            |   uint32_t {aka long unsigned int}
      |                                                            unsigned int
      |                                                           %lu

+++ lib/psgplay/sndh.c
@@ -285,7 +285,8 @@ static bool sndh_frames_subtag(struct sndh_cursor *cursor)
 			   (d[2] <<  8) |
 			    d[3];
 
-	snprintf(cursor->buffer, sizeof(cursor->buffer), "%u", t);
+	snprintf(cursor->buffer, sizeof(cursor->buffer), "%lu",
+			(long unsigned int)t);
 	tag_update(cursor->buffer, t, cursor);
 
 	cursor->offset += 4;
